requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
ale-py==0.8.0
astunparse==1.6.3
async-timeout==4.0.2
augmax
attrs==22.1.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.4.2
cached-property==1.5.2
cachetools==5.2.0
cffi==1.15.1
charset-normalizer==2.1.1
chex==0.1.5
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
cloudpickle==2.2.0
commonmark==0.9.1
contextlib2==21.6.0
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
dill==0.3.6
distrax==0.1.2
dm-tree==0.1.7
docker-pycreds==0.4.0
einops==0.6.0
etils==0.9.0
exceptiongroup==1.0.4
filelock==3.8.2
flatbuffers==22.12.6
flax==0.6.3
fonttools==4.38.0
freetype-py==2.3.0
frozenlist==1.3.3
fsspec==2022.2.0
ftfy==6.1.1
gast==0.5.3
gcsfs==2022.2.0
gitdb==4.0.10
GitPython==3.1.29
glcontext==2.3.7
glfw==1.12.0
google-api-core==2.11.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
gym==0.25.2
gym-notices==0.0.8
gym3==0.3.3
h5py==3.7.0
huggingface-hub==0.11.1
idna==3.4
imageio==2.22.4
imageio-ffmpeg==0.3.0
importlib-metadata==5.1.0
importlib-resources==5.10.1
iniconfig==1.1.1
Keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
ml-collections==0.1.1
moderngl==5.7.4
moviepy
msgpack==1.0.4
multidict==6.0.3
networkx==2.8.8
numpy==1.23.5
oauthlib==3.2.2
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
optax==0.1.4
orbax==0.0.23
packaging==22.0
pathtools==0.1.2
Pillow==9.3.0
pluggy==1.0.0
procgen==0.10.7
promise==2.3
protobuf==3.20.3
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyglet==2.0.1
Pygments==2.13.0
PyOpenGL==3.1.0
pyparsing==3.0.9
pyrender==0.1.45
pytest==7.2.0
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
scipy==1.9.3
sentry-sdk==1.11.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.28.0
tensorflow-probability==0.19.0
tensorstore==0.1.28
termcolor==2.1.1
tf-estimator-nightly==2.8.0.dev2021122109
tokenizers==0.12.1
tomli==2.0.1
toolz==0.12.0

--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.11.0+cu113
torchvision==0.12.0+cu113

tqdm==4.64.1
transformers==4.19.4
trimesh==3.17.1
typing_extensions==4.4.0
urllib3==1.26.13
wandb==0.13.6
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.2
zipp==3.11.0

opencv-python-headless
git+https://github.com/stepjam/PyRep.git
git+https://github.com/stepjam/RLBench.git
git+https://github.com/stepjam/YARR.git
pyrender

git+https://github.com/csmile-1006/procgenAISC.git
git+https://github.com/csmile-1006/procgen_highres_AISC.git
git+https://github.com/csmile-1006/procgen_highres.git